// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  genres: string[];
  starrings: [string[]];
  directors: string[]
  users: string[];
  emails: string[];
  avatars: string[];
  passwords: string[];
  previewVideoLinks: string[];
  videoLinks: string[];
  posterImages: string[];
  backgroundImages: string[];
  backgroundColors: string[];
};